add buttons and lcd

+++ _LegoSwerve/_Code/swerve/components/data_types/wheel_data.h
@@ -8,8 +8,8 @@ typedef struct {
     pid_ctrl_block_handle_t pid_ctrl;
     float target_wheel_rpm;
     float target_motor_rpm;
-    int target_angle;
-    int current_angle;
+    uint16_t target_angle;
+    uint16_t current_angle;
     float angle_error;
     float motor_rpm_differential;
